Updated build.xml with javadocs

+++ junit/org/homeunix/thecave/buddi/test/model/ScheduledTransactionTest.java
@@ -100,7 +100,7 @@ public class ScheduledTransactionTest {
 
 		d.updateAllBalances();
 
-		assertEquals(expected, a1.getBalance((Date) test));
+		assertEquals((long) expected, a1.getBalance((Date) test));
 	}
 
 }
